Restrict codes in kill_net_switch.sh. Reflect function can run.

+++ kill_nat_switch.sh
@@ -1,3 +1,8 @@
-PIDS=`ps aux | grep -E "bf_switchd.*nat" | sed "s/[^0-9]*\([0-9]*\).*/\1/g"`
-echo $PIDS
-kill -9 $PIDS > /dev/null 2>&1 #This will cause an error because it also kill the process of "grep", however it has terminated.
+PIDS=`ps aux | grep -E "[0-9]{2} bf_switchd.{10,}nat" | sed "s/[^0-9]*\([0-9]*\).*/\1/g"`
+if [ -n "$PIDS" ] # not empty
+then
+	echo "These running switch processes will be killed: "$PIDS
+	kill -9 $PIDS > /dev/null 2>&1 #This will cause an error because it also kill the process of "grep", however it has terminated.
+else
+	echo "No running switch process should be killed"
+fi
